feat(platform): added fuile enumerator and steam deck target for fun
