+++ src/linux.cpp
@@ -57,97 +57,6 @@ extern "C" void StopRedbookTrack()
     // TODO: implement CDROM features
 }
 
-struct TbFileFind {
-	std::vector<std::pair<std::string, std::string>> names;
-	size_t index = 0;
-};
-
-bool filespec_is_pattern(const char * filespec) {
-	return strchr(filespec, '*') != nullptr;
-}
-
-std::string directory_from_filespec(const char * filespec) {
-	const auto sep = strrchr(filespec, '/');
-	if (sep && sep != filespec) {
-		return std::string(filespec, sep - filespec);
-	} else {
-		return ".";
-	}
-}
-
-extern "C" TbFileFind * LbFileFindFirst(const char * filespec, TbFileEntry * fe)
-{
-	try {
-		auto ff = std::make_unique<TbFileFind>();
-		bool is_pattern = filespec_is_pattern(filespec);
-		std::string path;
-		if (is_pattern) {
-			path = directory_from_filespec(filespec);
-		} else {
-			path = filespec;
-		}
-		DIR *handle = opendir(path.c_str());
-		if (handle) {
-			while (true) {
-				auto de = readdir(handle);
-				if (!de) {
-					break;
-				}
-				if (strcmp(de->d_name, ".") == 0) {
-					continue;
-				}
-				if (strcmp(de->d_name, "..") == 0) {
-					continue;
-				}
-				const std::string file_path = path + "/" + de->d_name;
-				if (is_pattern) {
-					if (fnmatch(filespec, file_path.c_str(), FNM_FILE_NAME | FNM_CASEFOLD) != 0) {
-						continue;
-					}
-				}
-				struct stat sb;
-				if (stat(file_path.c_str(), &sb) < 0) {
-					continue;
-				}
-				if (!S_ISREG(sb.st_mode)) {
-					continue;
-				}
-				std::string key = de->d_name;
-				for (size_t i = 0; i < key.size(); i++) {
-					key[i] = (char)tolower((unsigned char)key[i]);
-				}
-				ff->names.emplace_back(key, de->d_name);
-			}
-			closedir(handle);
-		}
-		if (!ff->names.empty()) {
-			std::sort(ff->names.begin(), ff->names.end());
-			fe->Filename = ff->names[0].second.c_str();
-			return ff.release();
-		}
-	} catch (...) {}
-	return nullptr;
-}
-
-extern "C" int32_t LbFileFindNext(TbFileFind * ff, TbFileEntry * fe)
-{
-	try {
-		if (ff) {
-			ff->index++;
-			if (ff->index < ff->names.size()) {
-				fe->Filename = ff->names[ff->index].second.c_str();
-				return 1;
-			}
-		}
-	} catch (...) {}
-	return -1;
-}
-
-extern "C" void LbFileFindEnd(TbFileFind * ff)
-{
-	delete ff;
-}
-
 extern "C" int main(int argc, char *argv[]) {
 	return kfxmain(argc, argv);
 }
