+++ src/windows.cpp
@@ -77,56 +77,3 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance, LPSTR pCmdLine,
     LocalFree(szArglist);
     return kfxmain(argc, argv.data());
 }
-
-struct TbFileFind {
-    std::vector<std::pair<std::string, std::string>> names;
-    size_t index = 0;
-};
-
-extern "C" TbFileFind * LbFileFindFirst(const char * filespec, struct TbFileEntry * fentry)
-{
-    auto ffind = std::make_unique<TbFileFind>();
-    WIN32_FIND_DATA fd;
-    HANDLE handle = FindFirstFile(filespec, &fd);
-    if (handle == INVALID_HANDLE_VALUE) {
-        return nullptr;
-    }
-    do {
-        // Only return regular files. Skipping directories also drops "." and
-        // ".."; this matches the Linux implementation (which keeps S_ISREG
-        // entries only) so a plain "*" enumerates files on both platforms alike.
-        if (fd.dwFileAttributes & FILE_ATTRIBUTE_DIRECTORY) {
-            continue;
-        }
-        std::string key = fd.cFileName;
-        for (size_t i = 0; i < key.size(); i++) {
-            key[i] = (char)tolower((unsigned char)key[i]);
-        }
-        ffind->names.emplace_back(key, fd.cFileName);
-    } while (FindNextFile(handle, &fd));
-    FindClose(handle);
-    if (ffind->names.empty()) {
-        return nullptr;
-    }
-    std::sort(ffind->names.begin(), ffind->names.end());
-    fentry->Filename = ffind->names[0].second.c_str();
-    return ffind.release();
-}
-
-extern "C" int LbFileFindNext(struct TbFileFind * ffind, struct TbFileEntry * fentry)
-{
-    if (!ffind) {
-        return -1;
-    }
-    ffind->index++;
-    if (ffind->index >= ffind->names.size()) {
-        return -1;
-    }
-    fentry->Filename = ffind->names[ffind->index].second.c_str();
-    return 1;
-}
-
-extern "C" void LbFileFindEnd(struct TbFileFind * ffind)
-{
-    delete ffind;
-}
